+++ index.html
@@ -59,5 +59,5 @@
   <footer>
     <p>//</p>
   </footer>
-  <script src="./index.js"></script>
+  <script type="module" src="index.js"></script>
 </html>
