+++ modules/initializeBookList.js
@@ -1,0 +1,6 @@
+import BookList from '../index.js';
+
+export default function initializeBookList() {
+  const bookList = new BookList();
+  return bookList;
+}
